src/test/java/com/training/jpa/hibernate/jpahibernatepractice/repository/CourseRepositoryTest.java
package com.training.jpa.hibernate.jpahibernatepractice.repository;

import static org.junit.Assert.assertNull;
import static org.junit.jupiter.api.Assertions.assertEquals;

import javax.persistence.EntityManager;

import org.junit.jupiter.api.Test;
import org.junit.runner.RunWith;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.annotation.DirtiesContext;
import org.springframework.test.context.junit4.SpringRunner;
import org.springframework.transaction.annotation.Transactional;

import com.training.jpa.hibernate.jpahibernatepractice.JpaHibernatePracticeApplication;
import com.training.jpa.hibernate.jpahibernatepractice.entity.Course;
import com.training.jpa.hibernate.jpahibernatepractice.entity.Review;

@RunWith(SpringRunner.class)
@SpringBootTest(classes=JpaHibernatePracticeApplication.class)
class CourseRepositoryTest {

	private Logger logger = LoggerFactory.getLogger(this.getClass());
	
	@Autowired
	CourseRepository repository;
	
	@Autowired
	EntityManager em;
	
	@Test
	public void findById_basic() {
		Course course = repository.findById(10001L);
		assertEquals("JPA_TRAINING",course.getName());
	}
	
	@Test
	public void findById_firstLevelCacheDemo() {
		
		Course course = repository.findById(10001L);
		logger.info("First Course Retrieved {}", course);

		Course course1 = repository.findById(10001L);
		logger.info("First Course Retrieved again {}", course1);

		assertEquals("JPA in 50 Steps", course.getName());
		
		assertEquals("JPA in 50 Steps", course1.getName());
	}
	
	@Test
	@DirtiesContext //reset the data
	public void deleteById_basic() {
		repository.deleteById(10002L);
		assertNull(repository.findById(10002L));
	}
	
	@Test
	@DirtiesContext //reset the data
	public void save_basic() {
		//get course
		Course course = repository.findById(10001L);
		assertEquals("JPA_TRAINING",course.getName());
		
		//update
		course.setName("JPA_TRAINING_UPDATED");
		repository.save(course);
		
		//check
		Course course1 = repository.findById(10001L);
		assertEquals("JPA_TRAINING_UPDATED",course1.getName());
	}
	
	@Test
	@DirtiesContext //reset the data
	public void playWithEntityManager_basic() {
		repository.playWithEntityManager();
	}
	
	@Test
	@Transactional
	public void retrieveReviewsForCourse() {
		Course course = repository.findById(10001L);
		logger.info("{}",course.getReviews());
	}
	
	@Test
	@Transactional
	public void retrieveCourseForReview() {
		Review review = em.find(Review.class, 50001L);
		logger.info("{}",review.getCourse());
	}
	
	
	

}
